+++ CHANGELOG.md
@@ -18,6 +18,8 @@ PLEASE REMEMBER TO CHANGE THE '..main' WITH AN ACTUAL TAG in GITHUB LINK.
   * Moved `jax.experimental.stax` to `jax.example_libraries.stax`
   * Moved `jax.experimental.optimizers` to `jax.example_libraries.optimizers`
 
+* New features:
+  * Exposes `jax.scipy.linalg.qdwh` with major performance improvements.
 
 ## jax 0.2.24 (Oct 19, 2021)
 * [GitHub